requirements.txt
fastapi
groq
langchain-core
langchain-groq
pydantic
python-dotenv
python-socketio
requests
uvicorn
